refactor(replication): move stats contracts into crate
