+++ crates/replication/Cargo.toml
@@ -35,6 +35,7 @@ rustfs-utils = { workspace = true, features = ["http"] }
 s3s.workspace = true
 serde.workspace = true
 time.workspace = true
+tokio = { workspace = true, features = ["sync"] }
 url.workspace = true
 uuid = { workspace = true, features = ["v4", "serde"] }
